Fix the .not.rejectedWith tests.
In Chai `foo.should.not.throw(Constructor, 'string')` would fail only if
the error thrown is an instance of `Constructor` **and** if its message
includes `'string'`. An instance with a different message would not
cause a failure. An instance constructed by something else than
`Constructor` would not cause a failure.

.not.rejectedWith's semantics were incompatible with the above.

This commit fixes the test to work according to the correct semantics.